Move SamplingResult and MaskTarget into the modeling file
